Added "try" and "except" to handle errors.
I had to set the clientsocket to non-blocking mode with clientsocket.blocking(False)
because otherwise the program wouldn't continue past that point unless receiving a packet.
Since I sent the output of the clientsocket to a variable, I had to add "try" and "except"
to handle errors in case the variable ended up empty if there was no information in the socket.